User Profile Page Updated

+++ app/Http/Controllers/UserController.php
@@ -69,9 +69,9 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(User $user)
     {
-        return view('admin.users.show');
+        return view('admin.users.show', compact('user'));
     }
 
     /**
@@ -106,7 +106,9 @@ class UserController extends Controller
         $validated = $request->all();
 
         if ($request->file('avatar')) {
-            Storage::delete($user->avatar);
+            if ($user->avatar) {
+                Storage::delete($user->avatar);
+            }
             $validated['avatar'] = $request->file('avtar')->store('avatars');
         }
